Extract BookingCustomerFilterViewModel to a new file
